Update service and controller

+++ src/main/java/com/hrant/service/EmployeeService.java
@@ -4,7 +4,9 @@ import com.hrant.dto.DepartmentDto;
 import com.hrant.dto.EmployeeDto;
 import com.hrant.dto.PositionDto;
 import com.hrant.model.Employee;
+import com.hrant.repository.DepartmentRepository;
 import com.hrant.repository.EmployeeRepository;
+import com.hrant.repository.PositionRepository;
 import com.hrant.util.DtoConverter;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -19,14 +21,14 @@ public class EmployeeService {
     private final static Logger LOGGER = LoggerFactory.getLogger(EmployeeService.class);
 
     private final EmployeeRepository employeeRepository;
-    private final PositionService positionService;
-    private final DepartmentService departmentService;
+    private final PositionRepository positionRepository;
+    private final DepartmentRepository departmentRepository;
 
     @Autowired
-    public EmployeeService(EmployeeRepository employeeRepository, PositionService positionService, DepartmentService departmentService) {
+    public EmployeeService(EmployeeRepository employeeRepository, PositionRepository positionRepository, DepartmentRepository departmentRepository) {
         this.employeeRepository = employeeRepository;
-        this.positionService = positionService;
-        this.departmentService = departmentService;
+        this.positionRepository = positionRepository;
+        this.departmentRepository = departmentRepository;
     }
 
     public EmployeeDto addEmployee(EmployeeDto employeeDto) {
@@ -54,7 +56,15 @@ public class EmployeeService {
     }
 
     public void deleteEmployeeById(int id) {
+        EmployeeDto employeeDto = findEmployeeById(id);
+        if (employeeDto.getFName().isEmpty() && employeeDto.getLName().isEmpty()) {
+            LOGGER.info("The employee with the id " + id + " is not in the list");
+            System.out.println("The employee with the id " + id + " is not in the list");
+            return;
+        }
         employeeRepository.deleteById(id);
+        LOGGER.info("The employee with the id " + id + " is successfully deleted from the list");
+        System.out.println("The employee with the id " + id + " is successfully deleted from the list");
     }
 
     public EmployeeDto findEmployeeById(int id) {
@@ -65,8 +75,8 @@ public class EmployeeService {
         List<EmployeeDto> employeesDto = new ArrayList<>();
         for (Employee e : employees) {
             EmployeeDto employeeDto = DtoConverter.employeeToDto(e);
-            PositionDto positionDto = positionService.findPositionById(e.getPositionId()); //new PositionDto();
-            DepartmentDto departmentDto = departmentService.findDepartmentById(e.getDepartmentId()); //new DepartmentDto();
+            PositionDto positionDto = DtoConverter.positionToDto(positionRepository.findById(e.getPositionId()).orElse(DtoConverter.dtoToPosition(new PositionDto())));
+            DepartmentDto departmentDto = DtoConverter.departmentToDto(departmentRepository.findById(e.getDepartmentId()).orElse(DtoConverter.dtoToDepartment(new DepartmentDto())));
             employeeDto.setPositionName(positionDto.getName());
             employeeDto.setDepartmentName(departmentDto.getName());
             employeesDto.add(employeeDto);
